viewer: Trigger reloading texture images after a window resize.
SDL resets the OpenGL context when resizing a window, so all resources
(such as textures) get lost. This patch changes the viewer so that it
iterates over all textures and marks them as dirty whenever a resize
event occurs. This means during the next redraw, the images will get
reloaded (and the texture objects recreated).
This is a somewhat "dirty" fix and it only applies to textures, but
so far, this is the only thing that gets destroyed. If the renderer
is ever modified, this may have to be adjusted as well.